# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Park
  version: 1.0.0
  contact: {}
servers:
  - url: 'http://localhost:8089'
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      description: Login
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin
                password:
                  type: string
                  example: password
            examples:
              Login:
                value:
                  email: admin
                  password: password
      responses:
        '201':
          description: Success
          headers:
            Content-Length:
              schema:
                type: string
                example: '320'
            Date:
              schema:
                type: string
                example: Fri, 08 Mar 2024 09:17:47 GMT
            X-Request-Id:
              schema:
                type: string
                example: cnldfet4bsj6dg6a76b0
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      email:
                        type: string
                        example: admin@mail.com
                      token:
                        type: string
                        example: >-
                          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJhZG1pbkBtYWlsLmNvbSIsInJvbGUiOiJBRE1JTiIsImV4cCI6MTcxMDE0ODY2N30.zVRfQpjSuFqNEhsY8MjlBeMW47T_YpWyZr93oCpWhFY
                  message:
                    type: object
                    properties:
                      en:
                        type: string
                        example: Berhasil
                      id:
                        type: string
                        example: Sucessfully
                  success:
                    type: boolean
                    example: true
              examples:
                Success:
                  value:
                    data:
                      email: admin@mail.com
                      token: >-
                        eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJhZG1pbkBtYWlsLmNvbSIsInJvbGUiOiJBRE1JTiIsImV4cCI6MTcxMDE0ODY2N30.zVRfQpjSuFqNEhsY8MjlBeMW47T_YpWyZr93oCpWhFY
                    message:
                      en: Berhasil
                      id: Sucessfully
                    success: true
        '400':
          description: Bad Request
          headers:
            Content-Length:
              schema:
                type: string
                example: '167'
            Date:
              schema:
                type: string
                example: Fri, 08 Mar 2024 09:17:58 GMT
            X-Request-Id:
              schema:
                type: string
                example: cnldfhl4bsj6dg6a76bg
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 'email: admin does not validate as email'
                  message:
                    type: object
                    properties:
                      en:
                        type: string
                        example: Bad Request
                      id:
                        type: string
                        example: Bad Request
                  success:
                    type: boolean
                    example: false
              examples:
                Bad Request:
                  value:
                    error:
                      message: 'email: admin does not validate as email'
                    message:
                      en: Bad Request
                      id: Bad Request
                    success: false
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register
      description: Register
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user+2@mail.com
                password:
                  type: string
                  example: password
            examples:
              Register:
                value:
                  email: user+2@mail.com
                  password: password
      responses:
        '201':
          description: Success
          headers:
            Content-Length:
              schema:
                type: string
                example: '88'
            Date:
              schema:
                type: string
                example: Fri, 08 Mar 2024 15:11:01 GMT
            X-Request-Id:
              schema:
                type: string
                example: cnlil1d4bsj14li0tnig
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    properties:
                      en:
                        type: string
                        example: Berhasil
                      id:
                        type: string
                        example: Sucessfully
                  success:
                    type: boolean
                    example: true
              examples:
                Success:
                  value:
                    message:
                      en: Berhasil
                      id: Sucessfully
                    success: true
        '400':
          description: Email Already Registered
          headers:
            Content-Length:
              schema:
                type: string
                example: '152'
            Date:
              schema:
                type: string
                example: Fri, 08 Mar 2024 15:10:48 GMT
            X-Request-Id:
              schema:
                type: string
                example: cnliku54bsj14li0tni0
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: email already registered
                  message:
                    type: object
                    properties:
                      en:
                        type: string
                        example: Bad Request
                      id:
                        type: string
                        example: Bad Request
                  success:
                    type: boolean
                    example: false
              examples:
                Email Already Registered:
                  value:
                    error:
                      message: email already registered
                    message:
                      en: Bad Request
                      id: Bad Request
                    success: false
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...